test(ci): rename the live suite to contract and make it able to fail
