+++ e2e/contract-global-setup.ts
@@ -1,8 +1,8 @@
 /**
- * Fails the integration run immediately when it is misconfigured, rather than
+ * Fails the contract run immediately when it is misconfigured, rather than
  * letting every test skip itself and reporting green.
  *
- * `ZD_PROJECT_ID` is checked here — not in `scripts/check-integration-ran.mjs`
+ * `ZD_PROJECT_ID` is checked here — not in `scripts/check-contract-ran.mjs`
  * — for two reasons:
  *
  *   1. It is *always* our error. An unreachable KMS or a downed temp-email
@@ -20,12 +20,12 @@
 export default function setup(): void {
   if (!process.env.ZD_PROJECT_ID) {
     throw new Error(
-      'ZD_PROJECT_ID is not set, so every integration test would skip itself ' +
+      'ZD_PROJECT_ID is not set, so every contract test would skip itself ' +
         'and the run would falsely report success.\n\n' +
         'Local: copy `.env.example` to `.env` at the repo root and fill in ' +
         'ZD_PROJECT_ID.\n' +
         'CI: confirm the `ZD_PROJECT_ID` secret is populated and exposed to ' +
-        'the integration-test job.',
+        'the contract-test job.',
     )
   }
 }
