+++ e2e/contract/export-flow.test.ts
@@ -1,5 +1,5 @@
 /**
- * E2E integration test for wallet export.
+ * Contract test for wallet export.
  *
  * After OTP login, verifies that the SDK's export functions succeed against the
  * real backend / Turnkey and return well-formed encrypted export bundles:
