+++ e2e/contract/magic-link-flow.test.ts
@@ -1,16 +1,22 @@
 /**
- * E2E integration test for the OTP authentication flow.
+ * Contract test for the Magic Link authentication flow.
  *
- * Tests the complete OTP flow against a real KMS backend + Turnkey:
+ * Magic link is built on top of the OTP flow. Instead of sending a plain
+ * OTP code, Turnkey embeds the code into a clickable URL in the email.
+ * Whether a magic link (vs a plain code) is sent — and the link's URL
+ * template — is configured per-project on the backend
+ * (`wallet.otp_configs.magic_link_template`); the SDK just calls
+ * registerWithOTP. This test extracts the code from the magic-link URL when
+ * present and otherwise falls back to plain-code extraction.
+ *
+ * Flow:
  * 1. Create temp email account
- * 2. Register with OTP (sends email)
- * 3. Extract OTP code from email
+ * 2. Register with OTP
+ * 3. Extract OTP code from the magic link URL in the email (or plain code)
  * 4. Verify OTP with Auth Proxy
  * 5. Build client signature
  * 6. Login with OTP via backend
  * 7. Verify session works (whoami)
- *
- * Mirrors the Go E2E test at doorway-kms/testing/e2e/e2e_otp_test.go
  */
 
 import { beforeAll, describe, expect, it } from 'vitest'
@@ -28,7 +34,10 @@ import {
   EMAIL_POLL_TIMEOUT_MS,
   OTP_CODE_LENGTH,
 } from '../helpers/constants.js'
-import { extractOtpCode } from '../helpers/otp-utils.js'
+import {
+  extractOtpCode,
+  extractOtpCodeFromMagicLinkUrl,
+} from '../helpers/otp-utils.js'
 import {
   createNewAccount,
   ping,
@@ -37,13 +46,12 @@ import {
 import { createTestClient } from '../helpers/test-client.js'
 import { createTestStamper } from '../helpers/test-stamper.js'
 
-describe('OTP Authentication Flow', () => {
+describe('Magic Link Authentication Flow', () => {
   let projectId: string
   let authProxyConfigId: string
   let skipReason = ''
 
   beforeAll(async () => {
-    // Check backend availability first (fast fail)
     try {
       await waitForBackend(BACKEND_URL)
     } catch {
@@ -51,7 +59,6 @@ describe('OTP Authentication Flow', () => {
       return
     }
 
-    // Check email service availability
     try {
       await ping()
     } catch {
@@ -59,10 +66,8 @@ describe('OTP Authentication Flow', () => {
       return
     }
 
-    // Get auth proxy config ID from backend
     authProxyConfigId = await getAuthProxyConfigId(BACKEND_URL)
 
-    // Get project ID from env
     projectId = process.env.ZD_PROJECT_ID || ''
     if (!projectId) {
       skipReason = 'ZD_PROJECT_ID not set'
@@ -70,7 +75,8 @@ describe('OTP Authentication Flow', () => {
     }
   })
 
-  it('should complete the full OTP register + login flow', async (context) => {
+  it('should complete the full magic link register + login flow', async (context) => {
+    console.log(`Skipping magic link flow test: ${skipReason}`)
     context.skip(!!skipReason, skipReason)
 
     // Step 1: Create temp email account
@@ -78,16 +84,18 @@ describe('OTP Authentication Flow', () => {
     const email = emailAccount.address
     console.log(`Created temp email: ${email}`)
 
-    // Step 2: Create test stamper (Node.js ECDSA P-256 implementation)
+    // Step 2: Create test stamper
     const stamper = createTestStamper()
     const publicKey = await stamper.getPublicKey()
     expect(publicKey).toBeTruthy()
-    console.log(`Generated public key: ${publicKey!.substring(0, 16)}...`)
 
-    // Step 3: Create SDK client with test transport (includes Origin header)
+    // Step 3: Create SDK client
     const client = createTestClient(stamper)
 
-    // Step 4: Register with OTP (triggers email send)
+    // Step 4: Register with OTP. Whether the email carries a magic link or a
+    // plain code (and the link template) is configured per-project on the
+    // backend (`wallet.otp_configs.magic_link_template`); the client no longer
+    // supplies a template.
     const registerResult = await client.registerWithOTP({
       email,
       contact: { type: 'email', contact: email },
@@ -95,21 +103,29 @@ describe('OTP Authentication Flow', () => {
     })
     expect(registerResult.otpId).toBeTruthy()
     expect(registerResult.otpEncryptionTargetBundle).toBeTruthy()
-    console.log(`OTP initiated, otpId: ${registerResult.otpId}`)
+    console.log(`OTP initiated with magic link, otpId: ${registerResult.otpId}`)
 
-    // Step 5: Poll for email and extract OTP code
-    console.log('Waiting for OTP email...')
+    // Step 5: Poll for email and extract OTP code from magic link URL
+    console.log('Waiting for magic link email...')
     const emailContent = await searchForNewEmail(
       emailAccount.authToken,
       EMAIL_POLL_INTERVAL_MS,
       EMAIL_POLL_TIMEOUT_MS,
     )
-    const otpCode = extractOtpCode(emailContent, OTP_CODE_LENGTH)
+    console.log(`Email content preview: ${emailContent.substring(0, 200)}...`)
+
+    // Try extracting from URL first, fall back to plain OTP extraction
+    let otpCode = extractOtpCodeFromMagicLinkUrl(emailContent)
+    if (!otpCode) {
+      console.log(
+        'No magic link URL found in email, falling back to plain OTP extraction',
+      )
+      otpCode = extractOtpCode(emailContent, OTP_CODE_LENGTH)
+    }
     expect(otpCode).toBeTruthy()
     console.log(`Extracted OTP code: ${otpCode}`)
 
-    // Step 6: HPKE-seal the OTP attempt to the enclave's per-session target
-    // key, then verify with Auth Proxy.
+    // Step 6: HPKE-seal the OTP attempt and verify with Auth Proxy.
     const encryptedOtpBundle = await encryptOtpAttempt({
       otpCode: otpCode!,
       publicKey: publicKey!,
@@ -130,7 +146,7 @@ describe('OTP Authentication Flow', () => {
       stamper,
     })
     expect(clientSignature).toBeTruthy()
-    expect(clientSignature).toHaveLength(128) // 64 bytes = 128 hex chars
+    expect(clientSignature).toHaveLength(128)
     console.log('Client signature built')
 
     // Step 8: Login with OTP via backend
@@ -140,9 +156,9 @@ describe('OTP Authentication Flow', () => {
       projectId,
     })
     expect(loginResult.session).toBeTruthy()
-    console.log('OTP login successful')
+    console.log('Magic link login successful')
 
-    // Step 9: Parse and validate session
+    // Step 9: Validate session
     const session = parseSession(loginResult.session)
     expect(session.userId).toBeTruthy()
     expect(session.organizationId).toBeTruthy()
@@ -162,39 +178,5 @@ describe('OTP Authentication Flow', () => {
     console.log(
       `Whoami: userId=${whoami.userId}, orgId=${whoami.organizationId}`,
     )
-  })
-
-  it('should reject an invalid OTP code', async (context) => {
-    context.skip(!!skipReason, skipReason)
-
-    // Create temp email
-    const emailAccount = await createNewAccount()
-    const email = emailAccount.address
-
-    // Create stamper and client
-    const stamper = createTestStamper()
-    const publicKey = await stamper.getPublicKey()
-    const client = createTestClient(stamper)
-
-    // Register with OTP
-    const registerResult = await client.registerWithOTP({
-      email,
-      contact: { type: 'email', contact: email },
-      projectId,
-    })
-
-    // Try to verify with a wrong code
-    const wrongEncryptedBundle = await encryptOtpAttempt({
-      otpCode: 'WRONG12',
-      publicKey: publicKey!,
-      encryptionTargetBundle: registerResult.otpEncryptionTargetBundle,
-    })
-    const authProxyClient = createAuthProxyClient({ authProxyConfigId })
-    await expect(
-      authProxyClient.verifyOtp({
-        otpId: registerResult.otpId,
-        encryptedOtpBundle: wrongEncryptedBundle,
-      }),
-    ).rejects.toThrow()
   })
 })
