+++ e2e/contract/otp-flow.test.ts
@@ -1,22 +1,16 @@
 /**
- * E2E integration test for the Magic Link authentication flow.
+ * Contract test for the OTP authentication flow.
  *
- * Magic link is built on top of the OTP flow. Instead of sending a plain
- * OTP code, Turnkey embeds the code into a clickable URL in the email.
- * Whether a magic link (vs a plain code) is sent — and the link's URL
- * template — is configured per-project on the backend
- * (`wallet.otp_configs.magic_link_template`); the SDK just calls
- * registerWithOTP. This test extracts the code from the magic-link URL when
- * present and otherwise falls back to plain-code extraction.
- *
- * Flow:
+ * Tests the complete OTP flow against a real KMS backend + Turnkey:
  * 1. Create temp email account
- * 2. Register with OTP
- * 3. Extract OTP code from the magic link URL in the email (or plain code)
+ * 2. Register with OTP (sends email)
+ * 3. Extract OTP code from email
  * 4. Verify OTP with Auth Proxy
  * 5. Build client signature
  * 6. Login with OTP via backend
  * 7. Verify session works (whoami)
+ *
+ * Mirrors the Go E2E test at doorway-kms/testing/e2e/e2e_otp_test.go
  */
 
 import { beforeAll, describe, expect, it } from 'vitest'
@@ -34,10 +28,7 @@ import {
   EMAIL_POLL_TIMEOUT_MS,
   OTP_CODE_LENGTH,
 } from '../helpers/constants.js'
-import {
-  extractOtpCode,
-  extractOtpCodeFromMagicLinkUrl,
-} from '../helpers/otp-utils.js'
+import { extractOtpCode } from '../helpers/otp-utils.js'
 import {
   createNewAccount,
   ping,
@@ -46,12 +37,13 @@ import {
 import { createTestClient } from '../helpers/test-client.js'
 import { createTestStamper } from '../helpers/test-stamper.js'
 
-describe('Magic Link Authentication Flow', () => {
+describe('OTP Authentication Flow', () => {
   let projectId: string
   let authProxyConfigId: string
   let skipReason = ''
 
   beforeAll(async () => {
+    // Check backend availability first (fast fail)
     try {
       await waitForBackend(BACKEND_URL)
     } catch {
@@ -59,6 +51,7 @@ describe('Magic Link Authentication Flow', () => {
       return
     }
 
+    // Check email service availability
     try {
       await ping()
     } catch {
@@ -66,8 +59,10 @@ describe('Magic Link Authentication Flow', () => {
       return
     }
 
+    // Get auth proxy config ID from backend
     authProxyConfigId = await getAuthProxyConfigId(BACKEND_URL)
 
+    // Get project ID from env
     projectId = process.env.ZD_PROJECT_ID || ''
     if (!projectId) {
       skipReason = 'ZD_PROJECT_ID not set'
@@ -75,8 +70,7 @@ describe('Magic Link Authentication Flow', () => {
     }
   })
 
-  it('should complete the full magic link register + login flow', async (context) => {
-    console.log(`Skipping magic link flow test: ${skipReason}`)
+  it('should complete the full OTP register + login flow', async (context) => {
     context.skip(!!skipReason, skipReason)
 
     // Step 1: Create temp email account
@@ -84,18 +78,16 @@ describe('Magic Link Authentication Flow', () => {
     const email = emailAccount.address
     console.log(`Created temp email: ${email}`)
 
-    // Step 2: Create test stamper
+    // Step 2: Create test stamper (Node.js ECDSA P-256 implementation)
     const stamper = createTestStamper()
     const publicKey = await stamper.getPublicKey()
     expect(publicKey).toBeTruthy()
+    console.log(`Generated public key: ${publicKey!.substring(0, 16)}...`)
 
-    // Step 3: Create SDK client
+    // Step 3: Create SDK client with test transport (includes Origin header)
     const client = createTestClient(stamper)
 
-    // Step 4: Register with OTP. Whether the email carries a magic link or a
-    // plain code (and the link template) is configured per-project on the
-    // backend (`wallet.otp_configs.magic_link_template`); the client no longer
-    // supplies a template.
+    // Step 4: Register with OTP (triggers email send)
     const registerResult = await client.registerWithOTP({
       email,
       contact: { type: 'email', contact: email },
@@ -103,29 +95,21 @@ describe('Magic Link Authentication Flow', () => {
     })
     expect(registerResult.otpId).toBeTruthy()
     expect(registerResult.otpEncryptionTargetBundle).toBeTruthy()
-    console.log(`OTP initiated with magic link, otpId: ${registerResult.otpId}`)
+    console.log(`OTP initiated, otpId: ${registerResult.otpId}`)
 
-    // Step 5: Poll for email and extract OTP code from magic link URL
-    console.log('Waiting for magic link email...')
+    // Step 5: Poll for email and extract OTP code
+    console.log('Waiting for OTP email...')
     const emailContent = await searchForNewEmail(
       emailAccount.authToken,
       EMAIL_POLL_INTERVAL_MS,
       EMAIL_POLL_TIMEOUT_MS,
     )
-    console.log(`Email content preview: ${emailContent.substring(0, 200)}...`)
-
-    // Try extracting from URL first, fall back to plain OTP extraction
-    let otpCode = extractOtpCodeFromMagicLinkUrl(emailContent)
-    if (!otpCode) {
-      console.log(
-        'No magic link URL found in email, falling back to plain OTP extraction',
-      )
-      otpCode = extractOtpCode(emailContent, OTP_CODE_LENGTH)
-    }
+    const otpCode = extractOtpCode(emailContent, OTP_CODE_LENGTH)
     expect(otpCode).toBeTruthy()
     console.log(`Extracted OTP code: ${otpCode}`)
 
-    // Step 6: HPKE-seal the OTP attempt and verify with Auth Proxy.
+    // Step 6: HPKE-seal the OTP attempt to the enclave's per-session target
+    // key, then verify with Auth Proxy.
     const encryptedOtpBundle = await encryptOtpAttempt({
       otpCode: otpCode!,
       publicKey: publicKey!,
@@ -146,7 +130,7 @@ describe('Magic Link Authentication Flow', () => {
       stamper,
     })
     expect(clientSignature).toBeTruthy()
-    expect(clientSignature).toHaveLength(128)
+    expect(clientSignature).toHaveLength(128) // 64 bytes = 128 hex chars
     console.log('Client signature built')
 
     // Step 8: Login with OTP via backend
@@ -156,9 +140,9 @@ describe('Magic Link Authentication Flow', () => {
       projectId,
     })
     expect(loginResult.session).toBeTruthy()
-    console.log('Magic link login successful')
+    console.log('OTP login successful')
 
-    // Step 9: Validate session
+    // Step 9: Parse and validate session
     const session = parseSession(loginResult.session)
     expect(session.userId).toBeTruthy()
     expect(session.organizationId).toBeTruthy()
@@ -178,5 +162,39 @@ describe('Magic Link Authentication Flow', () => {
     console.log(
       `Whoami: userId=${whoami.userId}, orgId=${whoami.organizationId}`,
     )
+  })
+
+  it('should reject an invalid OTP code', async (context) => {
+    context.skip(!!skipReason, skipReason)
+
+    // Create temp email
+    const emailAccount = await createNewAccount()
+    const email = emailAccount.address
+
+    // Create stamper and client
+    const stamper = createTestStamper()
+    const publicKey = await stamper.getPublicKey()
+    const client = createTestClient(stamper)
+
+    // Register with OTP
+    const registerResult = await client.registerWithOTP({
+      email,
+      contact: { type: 'email', contact: email },
+      projectId,
+    })
+
+    // Try to verify with a wrong code
+    const wrongEncryptedBundle = await encryptOtpAttempt({
+      otpCode: 'WRONG12',
+      publicKey: publicKey!,
+      encryptionTargetBundle: registerResult.otpEncryptionTargetBundle,
+    })
+    const authProxyClient = createAuthProxyClient({ authProxyConfigId })
+    await expect(
+      authProxyClient.verifyOtp({
+        otpId: registerResult.otpId,
+        encryptedOtpBundle: wrongEncryptedBundle,
+      }),
+    ).rejects.toThrow()
   })
 })
