+++ package.json
@@ -20,9 +20,11 @@
     "test": "vitest run",
     "test:watch": "vitest",
     "test:coverage": "vitest run --coverage",
-    "test:integration": "vitest run --config e2e/vitest.integration.config.ts",
-    "test:integration:ci": "vitest run --config e2e/vitest.integration.config.ts --reporter=default --reporter=json --outputFile.json=integration-report.json",
-    "test:integration:check": "node scripts/check-integration-ran.mjs integration-report.json",
+    "test:integration": "vitest run --config vitest.integration.config.ts",
+    "test:integration:coverage": "vitest run --config vitest.integration.config.ts --coverage",
+    "test:contract": "vitest run --config e2e/vitest.contract.config.ts",
+    "test:contract:ci": "vitest run --config e2e/vitest.contract.config.ts --reporter=default --reporter=json --outputFile.json=contract-report.json",
+    "test:contract:check": "node scripts/check-contract-ran.mjs contract-report.json",
     "test:e2e": "playwright test --config e2e/playwright.config.ts",
     "test:e2e:headed": "playwright test --config e2e/playwright.config.ts --headed",
     "check:publishable": "node scripts/check-publishable.mjs",
